requirements.txt
playwright==1.51.0
openai>=1.0.0
chromadb>=0.4.0
anthropic>=0.20.0
pyyaml>=6.0
jinja2>=3.1.0
keyring>=24.0.0
requests>=2.31.0
